Desktop Video 11.5.1 SDK (05 May 2020) This SDK provides developer support for Desktop Video 11.5.1 that allows updating of hardware control and software interfaces for Desktop Video products.
Desktop Video 11.5.1 (05 May 2020)
This software update is for all customers using DeckLink 8K Pro, DeckLink Quad HDMI Recorder, DeckLink Duo 2, and DeckLink Quad 2 devices and addresses an occasional issue where captured audio may have appeared on incorrect channels. For DeckLink 8K Pro, compatibility is improved with SMPTE 425-1 when outputting 12-bit RGB 4:4:4 UltraHD and 4K DCI, and genlock and multichannel audio capture is now more reliable. DeckLink Quad HDMI Recorder has improved EDID support for capturing from game consoles and DSLR cameras, along with improved support for DVI video modes and more reliable sensing of input connections on startup.

+++ Mac/Samples/CaptureStills/Bgra32VideoFrame.h
@@ -39,7 +39,7 @@ private:
 	BMDFrameFlags			m_flags;
 	std::vector<uint8_t>	m_pixelBuffer;
 
-	std::atomic<uint32_t>	m_refCount;
+	std::atomic<ULONG>	m_refCount;
 
 public:
 	Bgra32VideoFrame(long width, long height, BMDFrameFlags flags);
